makes game guess number
This program makes the computer choose numbers from a certain distance.
Then it asks users to guess the number with given a clue. Instructions 
in the form of multiplication, division, and large and small against 
other numbers. If the users are wrong guess, the program will give 
users additional hints but lives users will decrease. The program will 
stop when the user has entered run out of life.
good luck!!!!